refactor: add file categories directory
