--- src/components/Navbar/index.tsx
@@ -1,14 +1,0 @@
-import type { FCWithLanyard } from "../../lib/types";
-import { DesktopNavbar } from "./DesktopNavbar";
-import { MobileNavbar } from "./MobileNavbar";
-
-export const Navbar: FCWithLanyard = ({ lanyard }) => {
-	return (
-		<>
-			{/* Created to block items from moving under navbar */}
-			<div className="fake-navbar" />
-			<MobileNavbar lanyard={lanyard} />
-			<DesktopNavbar lanyard={lanyard} />
-		</>
-	);
-};
