--- src/components/index.ts
@@ -1,5 +1,0 @@
-export * from "./Navbar";
-export * from "./Discord";
-export * from "./Button";
-export * from "./Tooltip";
-export * from "./Footer";
